Optimize tests execution.
