+++ sentry-jul/src/test/kotlin/io/sentry/jul/SentryHandlerTest.kt
@@ -6,7 +6,7 @@ import com.nhaarman.mockitokotlin2.verify
 import io.sentry.Sentry
 import io.sentry.SentryLevel
 import io.sentry.SentryOptions
-import io.sentry.test.checkEvent
+import io.sentry.checkEvent
 import io.sentry.transport.ITransport
 import java.time.Instant
 import java.time.LocalDateTime
