+++ sentry-logback/src/test/kotlin/io/sentry/logback/SentryAppenderTest.kt
@@ -12,7 +12,7 @@ import io.sentry.ITransportFactory
 import io.sentry.Sentry
 import io.sentry.SentryLevel
 import io.sentry.SentryOptions
-import io.sentry.test.checkEvent
+import io.sentry.checkEvent
 import io.sentry.transport.ITransport
 import java.time.Instant
 import java.time.LocalDateTime
