+++ sentry-spring/src/test/kotlin/io/sentry/spring/webflux/SentryWebfluxIntegrationTest.kt
@@ -11,7 +11,7 @@ import io.sentry.HubAdapter
 import io.sentry.IHub
 import io.sentry.ITransportFactory
 import io.sentry.Sentry
-import io.sentry.test.checkEvent
+import io.sentry.checkEvent
 import io.sentry.transport.ITransport
 import java.time.Duration
 import kotlin.test.assertEquals
